requirements.txt
fastapi
uvicorn
chromadb
openai
python-dotenv
pypdf
numpy
pandas
tiktoken
jinja2
